Use int64 outside Go int32 range

+++ tests/integration/cases/list_mixed_int_widths/Go_declaration_style_var@v1_18.go
@@ -1,7 +1,7 @@
 package main
 
 func main() {
-var my_data = []int{
+var my_data = []int64{
 	1,
 	1099511627776,
 }
